+++ src/api/deleteProfileImage.jsx
