+++ src/api/getStudyDetail.jsx
