+++ src/containers/JobPage/JobDetail/JobDetail.jsx
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import styled from "styled-components";
 
 import { useMutation, useQueryClient } from "react-query";
-import { GetJobDetail, JobDetailSelector, postComment } from "@/queries";
+import { GetJobDetail, JobDetailSelector, postComment } from "@/api";
 
 import { CardSmallSlider, CommentList, JobSideBar } from "@/containers";
 import { CommentInput, Toc } from "@/components";
